pyproject.toml
[project]
name = "cv-practice"
version = "0.1.0"
description = ""
authors = [
    {name = "ProfessionalSalmon",email = "terra.keerati@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.2,<3.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
